+++ src/ObjLoading/Game/T6/BSP/BSPCalculation.h
@@ -23,9 +23,9 @@ namespace BSP
     public:
         vec3_t min;
         vec3_t max;
-        int partitionIndex; // index of the partition the object is contained in
+        size_t partitionIndex; // index of the partition the object is contained in
 
-        BSPObject(float xMin, float yMin, float zMin, float xMax, float yMax, float zMax, int objPartitionIndex);
+        BSPObject(float xMin, float yMin, float zMin, float xMax, float yMax, float zMax, size_t objPartitionIndex);
     };
 
     class BSPLeaf
@@ -60,11 +60,11 @@ namespace BSP
         std::unique_ptr<BSPLeaf> leaf;
         std::unique_ptr<BSPNode> node;
 
-        int level; // level in the BSP tree
+        size_t level; // level in the BSP tree
         vec3_t min;
         vec3_t max;
 
-        BSPTree(float xMin, float yMin, float zMin, float xMax, float yMax, float zMax, int treeLevel);
+        BSPTree(float xMin, float yMin, float zMin, float xMax, float yMax, float zMax, size_t treeLevel);
         void splitTree();
         void addObjectToTree(std::shared_ptr<BSPObject> object);
     };
